Modify new init migration to the original integer gpu_count and add new migration file with float conversion

+++ src/sc_crawler/alembic/versions/dad8a1f0f455_v2_0_0_migration_reset.py
@@ -627,7 +627,7 @@ def create_server_table(is_scd: bool):
         ),
         sa.Column(
             "gpu_count",
-            sa.Float(),
+            sa.Integer(),
             nullable=False,
             comment="Number of GPU accelerator(s).",
         ),
